AI-bot/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.10.0
WTForms==3.1.1
Werkzeug==3.0.1
requests==2.31.0
python-dotenv==1.0.0
Pillow>=10.2.0
authlib==1.3.0
pytz==2024.1
urllib3>=1.26.0
gigachat>=0.1.0

